Fix second batch of MSVC compilation errors
- Add #include <vector> and <cstdint> to HwpFormat.h/cpp, DocFormat.cpp, HtmlFormat.cpp
  (fixes std::vector not found, WideCharToMultiByte/WriteFile arg count errors)
- Fix SplashScreen: UINT_PTR timerId (was UINT, narrowing), pass HDC from
  BeginPaint into OnPaint to avoid unused-variable warning and redundant GetDC
- Remove unused holeR variable in DrawLogoIcon
- Fix std::max(0, LONG) type mismatch → std::max(0L, LONG) in Editor.cpp
- Replace undefined TB_GETBOUNDINGRECT with TB_GETMAXSIZE in MainWindow.cpp
- Add explicit BYTE casts on TBBUTTON fsState/fsStyle fields to suppress
  narrowing warnings; fix separator width (iBitmap=276, not dwData=276)

+++ src/Editor.cpp
@@ -344,7 +344,7 @@ void Editor::Indent(int direction) {
     fmt.cbSize = sizeof(fmt);
     SendMessageW(m_hwnd, EM_GETPARAFORMAT, 0, reinterpret_cast<LPARAM>(&fmt));
     fmt.dwMask = PFM_STARTINDENT;
-    fmt.dxStartIndent = std::max(0, fmt.dxStartIndent + direction * 720);
+    fmt.dxStartIndent = std::max(0L, fmt.dxStartIndent + static_cast<LONG>(direction * 720));
     SendMessageW(m_hwnd, EM_SETPARAFORMAT, 0, reinterpret_cast<LPARAM>(&fmt));
 }
